@@ -2,19 +2,22 @@
   "name": "workfront-cookie",
   "version": "3.3.1",
   "description": "Extracts values from workfront cookies",
-  "main": "dist/workfront-cookie.min.js",
+  "main": "dist/workfront-cookie.umd.js",
+  "module": "dist/workfront-cookie.esm.js",
   "files": [
     "dist",
-    "index.d.ts"
+    "typings"
   ],
+  "types": "./typings/index.d.ts",
   "scripts": {
-    "prepublish": "npm run build",
-    "build": "webpack",
-    "dev": "webpack --watch",
-    "lint": "eslint --ext .js .",
-    "test:lint": "npm run lint",
-    "test:sources": "karma start karma.conf.js",
-    "test:typings": "./node_modules/.bin/_mocha -- test/typings.spec.js",
+    "build": "run-p build:*",
+    "build:bundles": "rollup --config rollup.config.js",
+    "build:bundles-minified": "rollup --config rollup.config.min.js",
+    "build:typings": "tsc --declaration --emitDeclarationOnly --noEmitOnError --declarationDir typings",
+    "prettier": "prettier --write \"**/*.{ts,tsx,js,jsx,css,md,yml}\"",
+    "test:typecheck": "tsc --noEmit",
+    "test:sources": "jest --coverage",
+    "test:sources:watch": "jest --watch",
     "test": "run-p test:*",
     "coverage": "cat coverage/*/lcov.info | codecov"
   },
@@ -38,24 +41,20 @@
   },
   "homepage": "https://github.com/Workfront/workfront-cookie#readme",
   "devDependencies": {
+    "@commitlint/cli": "7.5.2",
+    "@commitlint/config-conventional": "7.5.0",
+    "@types/jest": "24.0.11",
     "codecov": "3.2.0",
-    "eslint": "5.15.3",
-    "eslint-loader": "2.1.2",
-    "expect.js": "0.3.1",
-    "istanbul-instrumenter-loader": "3.0.1",
-    "karma": "4.0.1",
-    "karma-cli": "2.0.0",
-    "karma-coverage": "1.1.2",
-    "karma-mocha": "1.3.0",
-    "karma-phantomjs-launcher": "1.0.4",
-    "karma-sourcemap-loader": "0.3.7",
-    "karma-webpack": "3.0.5",
-    "mocha": "5.2.0",
+    "husky": "1.3.1",
+    "jest": "24.5.0",
+    "lint-staged": "8.1.5",
     "npm-run-all": "4.1.5",
-    "phantomjs-prebuilt": "2.1.16",
-    "typescript": "3.3.4000",
-    "typescript-definition-tester": "0.0.6",
-    "webpack": "4.29.6",
-    "webpack-cli": "3.3.0"
-  }
+    "prettier": "1.16.4",
+    "rollup": "1.7.4",
+    "rollup-plugin-typescript": "1.0.1",
+    "rollup-plugin-uglify": "6.0.2",
+    "ts-jest": "24.0.0",
+    "typescript": "3.4.1"
+  },
+  "dependencies": {}
 }
