@@ -14,14 +14,10 @@
  * limitations under the License.
  */
 
-var readCookie = require('./readCookie')
-var cookieName = require('./cookieName')
-
-function getSessionID(cookieString) {
-    var cookie = readCookie(cookieName, cookieString)
-    if (cookie) {
-        return cookie.split('#')[0]
-    }
+export function readCookie(name: string, cookieString?: string): string | undefined {
+  cookieString = cookieString || document.cookie
+  const match = cookieString.match(new RegExp('(^|;\\s*)(' + name + ')=([^;]*)'))
+  if (match) {
+    return decodeURIComponent(match[3])
+  }
 }
-
-module.exports = getSessionID
